Packet Tracer 1
This activity was a huge eye opener, containing the basics of introducing new devices such as the backup router and using the terminal from a connected device to the router via a USB port, to configure the name. As well as this, I learnt the differences between the topologies, such as the logical and physical views. Being able to distinguish between the two is pivotal for understanding the layouts.